# Remove util.premultiply