--- a/js/render/draw_fill.js
+++ b/js/render/draw_fill.js
@@ -1,6 +1,5 @@
 'use strict';
 
-var util = require('../util/util');
 var pixelsToTileUnits = require('../source/pixels_to_tile_units');
 
 module.exports = draw;
@@ -9,7 +8,7 @@ function draw(painter, source, layer, coords) {
     var gl = painter.gl;
     gl.enable(gl.STENCIL_TEST);
 
-    var color = util.premultiply(layer.paint['fill-color']);
+    var color = layer.paint['fill-color'];
     var image = layer.paint['fill-pattern'];
     var opacity = layer.paint['fill-opacity'];
     var isOutlineColorDefined = layer.getPaintProperty('fill-outline-color');