GL JS had the notion of "premultiplying" colors just before rendering them, transforming rgba arrays into the color format expected by GL (all components between 0 and 1, all components multiplied by the alpha value).

This PR removes the premultiply step, opting instead to format the colors in the format expected by GL end-to-end in order to

 - reduce the complexity of GL JS
 - get minor perf gains in regular rendering
 - get big perf gains in data-driven styling rendering

### Frame Duration Benchmark

**this branch** 9.9 ms per frame. 4% of frames took longer than 16ms.
**master** 10.3 ms per frame. 5% of frames took longer than 16ms.

cc @mollymerp @jfirebaugh @mourner