@mollymerp 2016-06-07
LGTM – seems like this simplifies the code base and does not have any perf drawbacks. I ran the frame duration benchmarks locally and the results were almost identical: both had 5.8ms / frame and 0% taking longer than 16 ms. 

🚢