@mourner 2016-06-07
@ansis wasn't the original reason to premultiply in the final stage something to do with interpolating between two semi-transparent colors + opacity? I have a feeling this breaks something in a subtle way not catched by the test suite...