@lucaswoj 2016-06-07
You might be thinking of the behaviour that was removed in https://github.com/mapbox/mapbox-gl-js/pull/1347. 

It's possible that there is an edge case that occurs in transitions between semi-transparent colors but everything looked 👍 in my testing.